[SSD-69] feat: 회원가입 API Main 충돌문제 해결 및 코드 수정
