+++ backend/src/main/java/com/timepaper/backend/global/auth/dto/EmailCertificationRequestDto.java
@@ -1,4 +1,4 @@
-package com.timepaper.backend.domain.auth.dto;
+package com.timepaper.backend.global.auth.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
