+++ backend/src/main/java/com/timepaper/backend/global/auth/service/AuthService.java
@@ -1,16 +1,25 @@
 package com.timepaper.backend.global.auth.service;
 
+import com.timepaper.backend.domain.javaemail.JavaEmailSender;
+import com.timepaper.backend.domain.user.entity.User;
+import com.timepaper.backend.global.auth.dto.CertificationNumberRequestDto;
+import com.timepaper.backend.global.auth.dto.EmailCertificationRequestDto;
+import com.timepaper.backend.global.auth.dto.SignupDto;
+import com.timepaper.backend.global.auth.repository.AuthRepository;
 import com.timepaper.backend.global.auth.token.service.RefreshTokenService;
 import com.timepaper.backend.global.auth.token.util.JWTUtil;
 import com.timepaper.backend.global.auth.token.util.RefreshTokenUtil;
 import jakarta.servlet.http.HttpServletResponse;
 import java.time.Duration;
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseCookie;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Service
@@ -20,6 +29,9 @@ public class AuthService {
   private final RefreshTokenService refreshTokenService;
   private final JWTUtil jwtUtil;
   private final RefreshTokenUtil refreshTokenUtil;
+  private final AuthRepository authRepository;
+  private final RedisTemplate<String, String> redisTemplate;
+  private final JavaEmailSender javaEmailSender;
 
   public void setTokensResponse(HttpServletResponse response, Authentication authentication) {
     String accessToken = jwtUtil.createToken(authentication);
@@ -64,5 +76,42 @@ public class AuthService {
         .build();
   }
 
+
+  public boolean emailverification(EmailCertificationRequestDto dto) {
+    boolean emailcheck = authRepository.findByEmail(dto.getEmail());
+
+    if (emailcheck) {
+      return false;
+    } else {
+      String randomCode = UUID.randomUUID().toString().replace("-", "").substring(0, 6)
+          .toUpperCase();
+      redisTemplate.opsForValue().set(dto.getEmail(), randomCode, Duration.ofMinutes(5));
+
+      System.out.println(randomCode); // 인증번호
+
+//      JavaEmailDto javaEmailDto = new JavaEmailDto(dto.getEmail(), "TimePaper", randomCode);
+//
+//      javaEmailSender.sendJavaEmail(javaEmailDto);
+      return true;
+    }
+  }
+
+  @Transactional
+  public boolean checkEmailVerificationCode(CertificationNumberRequestDto dto) {
+    String randomCode = redisTemplate.opsForValue().get(dto.getEmail());
+    System.out.println(randomCode);
+    return randomCode != null && randomCode.equals(dto.getCheckNum());
+  }
+
+  @Transactional
+  public boolean signUp(SignupDto dto) {
+    try {
+      User auth = dto.toEntity(dto);
+      authRepository.save(auth);
+      return true;
+    } catch (Exception e) {
+      return false;
+    }
+  }
 
 }
