+++ backend/src/main/resources/application.properties
@@ -4,7 +4,7 @@ spring.datasource.username=${DATABASE_USERNAME}
 spring.datasource.password=${DATABASE_PASSWORD}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=update
+spring.jpa.hibernate.ddl-auto=create-drop
 spring.docker.compose.lifecycle-management=start-and-stop
 logging.level.org.springframework.security=DEBUG
 jwt.secret=${JWT_SECRET}
